Make `Factory` metaclass general and move it to `utils`
- SingletonFactory inherits from AbstractSingletonType and Factory:
to be used for creating singleton's out of an interface.
- `types` attribute of Factory classes (like `Database`) is dynamically
populated by using **glob** to find implementation files and import
them, and then using `cls.__base__.__subclasses__()` to create list with
the visible names that are subclassing cls.__base__ (i.e. `AbstractDB`
for `Database`).
- Invoking with arguments after the first call on a SingletonFactory
class would result in raising a ValueError.