Fix experiment trial selection race condition
Why:

Concurrent workers could reserve the same trial with current
implementation. We need to make sure a process will abort reserving a
trial if another one already did so.

How:

Replace `db.write` by `db.read_and_write()` and use not only the trial's
id but also its status for the query. Doing so, a change of status between
`db.read()` and `db.read_and_write()` will abort the write and force the
process to rollback and try a reservation on another trial.