Add method ensure_index to databases
Why:

Our main and only database back end is a NoSQL one and only `_id` is
indexed by default. That means every `read` operation (including reads
necessary in updates with queries) needs to parse the entire collection
unless the query specifies the _id. This is highly inefficient. Adding
indexes on keys which are often used in queries will significantly
improve the database operations.

Additionally, we want to force some keys to be unique, namely the
tuple (experiment name, user name). To do so we need a method to specify
unique indexes.

How:

Add `ensure_index` method to `AbstractDB`.

- Keys can be specified by name and optionally by sort order too.
- Keys can be forced to be unique.
- Generic sort orders are added to `AbstractDB`.
- `MongoDB` backend uses `dbcollection.create_index()`