Remove experiment duplicate name from unittests
Why:

Unittest configuration contained two experiments with the same tuple
(name, user). The addition of enforced uniqueness on this tuple makes
many unittest crash.

How:

Change the name of the second experiment with duplicate keys and adapt
unittest to either test one or the other. For unit tests requiring many
instances, change the query to (name, ) instead of (name, user) so that
the output is still more than one experiment.