Rebuild experiment in case of race condition
Why:

The configuration of the experiment is not atomic and race conditions
are likely to happen when many workers are submitted at once to start
optimization.

How:

After all configuration sources are merged, experiment configuration
attempts to write it to DB. If write fails because of a race condition,
`cli` will rebuild the experiment from scratch. Doing this will trigger
a fresh read from DB and load the configuration saved by the _winning_
worker.