Use namespace pymongo to access MongoClient
Why:

When importing using `from pymongo import MongoClient`, we are assigning
`MongoClient` to the local namespace, making it different than
`pymongo.MongoClient`. It makes the mocking the of object more tedious
since we now need to access it via
`metaopt.core.io.database.mongodb.MongoClient` rather than
`pymongo.MongoClient`. Furthermore, the use of `pymongo.MongoClient` is
preferable because it is more clean than `MongoClient`. It is clear that
it is part of `pymongo` and not a local custom wrapper we defined.