Client: Add helper function for manual insertion of parameter sets
Why:
A user might want to populate a collection of trials with one's own
choices of set of parameters for a particular experiment. Call it
intuition, call it bias, call it experience; the user should be able to
act as if one was the Producer itself.

How:
User can either write a second script of one's own, where some tuples of
appropriate sets of parameters are calculated and fed to
`metaopt.client.insert_trials`, or one can do the same thing from
iPython - for example. The function is written as part of the
`metaopt.client` interface to facilitate this interaction.

In order to work, an `Experiment` object must be instantiated, and this
need in turn a `Database` object to lie in the Python process. So user
is asked to pass necessarily the experiment's name which the trials will
be associated with, and take care of a way to discover the Database. The
latter will be done in the same manner that `metaopt.core.cli.infer_experiment`
discovers the `Database`. That is either by default locations,
environmental variables, or an external configuration. The last one is
enabled by allowing an extra optional keyword argument in the helper
function, ``cmdconfig``, which is expected to have the key 'database'
which is a dictionary with configuration details, as usual.

Note:
The experiment must exist already as a configuration in the specified
database. **So, currently pre-population of an experiment's new trials is
not supported**.

Finally:
All points suggested by the user are checked against the experiment's
parameter space. In case of error, an optional argument controls whether
an `AssertionError` will be raised or the failing parameter sets will be
ignored.

Squash with Client