Add ``requires`` class attribute to BaseAlgorithm
Why:
This is a way for an algorithm's implementation to request
a particular form of the Space object, which is necessary for it to
operate. This will be used by `metaopt.core.worker.primary_algo` to
define the appropriate transformation of the `Space` and the points
in the parameter space.