Add default value support for dimensions
Why: Default values are necessary when inserting a trial manually and
giving only a subset of the space of an experiment as fixed dimensions.
When a dimension is missing, it'll go look inside the configuration and
see if the default value has been set, if not, it'll stop.

How: By adding the default value as a named parameters of the distribution,
the dimension can check if it belongs to its domain and cast it to the correct
type. This limits the actual code impact since the parsing of the parameters
was already done inside the _build function of DimensionBuilder.