Modularize commandline commands parsing and execution
Why: The current way of parsing and executing commandline commands
like 'hunt' and 'insert' relies on modifying OrionArgsParser by adding
multiple functions, which creates a lot of duplicate and useless
boilerplate code. It also means that the users need to fidget with
Orion's source code to extend Orion's behavior with new commands.
This refactoring aims at modularizing the commands' parsing and
execution by delegating the job to the module running the command,
which will let the users extend Orion as they wish, easily. They
just need to follow the current API by providing a 'get_parser'
function and making sure that the 'func' key is set to the function
to call once the args are parsed.

How: This modification relies on mutliple things. First, the cli
__init__.py obtains the path to its current directory, then loops
through all the python files inside that directory. If a python script
provides a 'get_parser' function, it is call, passing the
OrionArgsParser instance as a parameter, so that the module can had
its own argument.

This change should not impact the code or current test suites, since
the command execution is exactly the same.