Add `insert` command to generate a single trial
Why: The `insert` command let's the user insert a single trial
with specific values. This is useful in cases where the hyperparameters
are approximately known or if the user is testing an implementation
of someone else's model and the hyperparameters are known.

How: The user gives the different point through the command line
following closely the same convention as for priors' distibution
by giving the namespace then its value like this : "-x=1". The
insert command then makes sure of multiple things:

    -It first checks that there is a default value for all the
    dimensions inside the experiment's space that are not specified
    by the user.

    -Then it checks that there is no namespace outside of the
    experiment space.

    -Finally, it makes sure that the given value lies inside the
    prior's distribution interval

If these conditions are respected, a new trial is inserted
inside the database.