Add type validation for Dimension inside `contains`
Why: There is a bug (or a feature, not sure) inside numpy that
would make comparing two arrays raise an error because one of
the values was of type `str` since it came directly from the command
line. This modification casts the point being check to the dimension's
type to make sure all the values are valid.