Change non-categorical Dimensions __contains__ to handle strings
Why: Currently, checking if "1" is inside a Real or Integer dimension's
interval will result in a numpy `TypeError` since it is trying to
compare two different type of ndarray (one as string and the other as
the dimension interval's type).
This change gets rid of the base class `Dimension` __contains__
function and places the responsability of checking if a point is
valid inside their interval on the actual implementations. This
doesn't have any impact on actual code since `Dimension` objects
should not be instanciated in the first place

How: Integer dimensions will use numpy.asarray with dtype = float64,
which solves issues with values passed as real (such as `1.0`) while
keeping the behavior the same since the actual check for being an
integer is inside the if.