Make sure inserted trials use the correct type
Why: Since the command line arguments are all strings, there needs
to be a way to verify that its value is contained inside a Dimension.
This change combines a typical check with `not in` as well as an eval
call to obtain the actual representation of that arg.
The validate function handles the case where the string "None" is
inside a categorical dimension's categories.