Refactored ExperimentBranchBuilder and BranchingPrompt
Why: the old implementation basically sucked. Maintaining two different
dictionaries, as well as multiple lists of names that needed to be repo
pulated, it was way too painful and prone to bug.
This new implementation stores everything inside a `Conflict` object
and maintains only a list of them, which offers easy mapping and filter
ing.