Add adapters for experiment version control
Why:

Experiment branches because of changes in the configuration or
the user's code. This make experiments incompatible with one another
unless we define adapters such that a branched experiment B can access
trials from parent experiment A.

How:

Adapters have two main methods, forward and backward. The method forward
defines how trials from parent experiment A are filtered or adapted to
be compatible with experiment B. Modifications are only applied at
execution time and are not saved anywhere in the database. Adapters only
provides a *view* on an experiment.

There is adapters for

* Dimension addition
* Dimension deletion
* Dimension renaming
* Change of dimension prior
* Change of algorithm
* Change of code
* Combining different adapters

Adapters all have a `to_dict` method which provides sufficient
information to rebuild the adapter. This is to facilitate save of
adapters in a database and retrieval.

Adapters can be build using the factory class Adapter(**kwargs) or using
Adapter.build(list_of_dicts).