Add tree data structure for EVC
Why:

Experiment version control requires building trees of the experiments so
that we can fetch trials from one experiment to another or navigate from
one experiment to another to visualise different statistics.

How:

Implement a simple TreeNode and tree iterators to support the tree data
structure of the experiment version control. TreeNode is a generic class
which can carry arbitrary python objects. It comes we basic methods to
set parent and children. A method `map` allows to apply functions
recursively on the tree in a generic manner.