Extend Factory to find subclasses recursively
Why:

If a subclass of OptimizationAlgorithm is extended, the meta-class
Factory will not find it's implementation. Let's say for instance that
we define a global class ScikitOptimizationAlgorithm which integrates
generalities for all different algorithms from scikit-optimize, then all
children of ScikitOptimizationAlgorithm would not be found by the
Factory unless they are specifically registered by entry points.

How:

Call recursively on cls.__subclasses__() and list them all.