Fix Param and Result __slots__ attribute
Why:

Classes do not inherit __slots__, it needs to be redefined in each
children. The parent class Value's __init__ depends on the combination
of kwargs parsing and __slots__ to raise AttributeError on invalid
arguments. Since Param and Result do not inherint __slots__, they fail
to raise an AttributeError when passed bad attributes.