Add ExperimentView wrapper
Why:

With experiment version control there is the possibility of having many
experiments instantiated at the same time but only one should be
editable, the current one. All the others should not be able to make
modifications in the database. This commit is an attempt at reducing
write errors by wrapping the experiment with a read-only wrapper which
defines which attributes can be accessed and which not.

Note:

It will still be possible to access the write methods of the experiment,
by explicitly fetching it with by accessing the sub attribute of the
wrapper (ex: `experiment_view._experiment.register_trials()`).
Nonetheless the current wrapper makes it less likely to do so by error.

How:

Create an empty class with a `valid_attributes`. list which defines which
attribute and method may be accessed. The method `.__getattr__` filters out
the request on the wrapper based on `valid_attributes`.

The property `is_done` is slightly modified in the view since
`experiment.is_done` would normally do a write on the database. If
`wrapper.is_done` returns True, it will not update the database like the
normal experiment would.