Refactor experiment build into ExperimentBuilder
Why:

The instantiation of an `Experiment` is not a trivial process when the
user request an experiment with specific options. One can easily create
a new experiment with `ExperimentView('some_experiment_name')`, but the
configuration of a _writable_ experiment is less straighforward. This is
because there is many sources of configuration and they have a strict
hierarchy. This commit attempts to unify the process of building an
experiment into a single module/class.

How:

The builder can build an experiment only provided the command args. All
fetch of other source of configuration are executed internally. The
implementation is less efficient to favor simplicity. Yet the loss of
efficiency should not impact the global execution as creating an
experiment happens rarely and requires only a few query on the
database. This is also true in the context of experiment version control
trees.

The builder is built as a class such that the methods' efficiency could
be improved latter on by caching configuration fetches to avoid
duplicate fetches.

All cli commands are adapted to use the new experiment builder.