Auto configure experiment view
Why:

The experiment's space is only instantiated at configuration time. The
space is an attribute which may be very important in experiment views,
so self-configuring the view is primordial.

How:

Use the experiment's configuration in database to self configure the
experiment. If the experiment is not in database, the experiment view
will raise a ValueError at instantiation time.
orion/core/worker/experiment.py