Do no set experiment on pending by default
Why:

If an experiment is broken, it should not be automatically resumed
simply by fetching the experiment from database and configuring it.
Also, if the experiment is done and the termination criterion does not
change, then it should not be set to pending automatically.

How:

At configuration time, test if status is None or status is not broken
and the termination criterion is False. If yes, then set status to
pending.