Add fixture to set proper ids in the db
Why:

The trials are using experiment names as ids in the unit-tests
configuration for convenience, but with new `refers` structure the
experiment document will do so as well. We need to update those
references to the actual experiment ids.

How:

Upon database instantiation, fetch all experiments and update the ids by
matching the names.

Note:

This is only done with user "tsirif".