Add experiment node mechanisms
Why:

The tree of experiments have common use cases like fetching trials and
lazy experiment instantiation. Defining a child `ExperimentNode` is
convenient to group such operations.

How:

ExperimentNode carries an experiment in `item`. If node is created using
only the experiment's name, it will instantiate the experiment only upon
access to `item`. `ExperimentNode.fetch_trials` returns a list of trials
fetched recursively in the EVC tree.

Helper functions to apply operations on the tree, not the node
specifically, are kept outside `ExperimentNode`. Results of helper
functions on the EVC are returned in the form of a tree to keep the
original structure (can be useful when analysing an EVC tree by keeping
experiment results separated in each node).