Add EVCBuilder
Why:

Building the EVC tree can by tricky if we support trimming for
optimization on smaller EVC subtree. For example user can define a root
and some leafs that should be the extremums of the tree. Those can be
different than the actual root and leafs of the global EVC tree, making
the trimmed version a small subset of the global version.

How:

Wraps ExperimentBuilder fonctionalities to parse command line.

The process of building the tree is the following:

1. Fetch all documents with a given root_id (root of global tree)
2. Build the global tree based on those shallow documents
   (only minimal info fetched from db)
3. Trim the tree based on root name and leaf names specified by user
4. Turn the shallow nodes into experiment nodes (lazy initialization)

The experiment is then connected to the EVC tree, before being returned.