Use algorithms.is_done in view.is_done
Why:

In first implementation, the view was wrapped a non-configured
experiment, hence the algorithm was not instantiated. It turns out
wrapping a non configured experiment was not very good since
experiment.space and experiment.algorithms have useful information even
in _view-mode_.

Now that the view wraps a fully configured experiment, the property
`is_done` should uses the instantiated algorithm.