Fix children reinsertion ordering
Why:

In `TreeNode`, the method `add_children` applied on a node that is
already the child of the current parent would remove it from the list of
child and add it back to the end of it. It should stay at the same
position.

How:

Only drop child's parent if it is to be added as a child. Is it is
already a child, do nothing.