Remove status attribute from Experiment
Why:

The status is not useful since properties is_done and
is_broken (coming soon) can be used to verify the state
of the experiment. As the status can be inferred, and experiments
are not very likely to be queried based on status (in opposition
to trials which are often queried based on status), there is
no reason to maintain a status attribute.

How:

Replace tests on status (which all had the function of testing if an
experiment is configured or found in DB) with test on _id.