Add id property to experiment
Why:

The status of experiments was used to test if the experiment
was configured or not. Now that the status attribute does not exist
anymore, the id can be used in replacement, so it needed a public
getter attribute.

Note:
Maybe experiment.is_configured would be better for such tests?