Move datetime setup in experiment
Why:

Most metadata is meant to trigger experiment branching events, but
not the datetime which is when the experiment is registered (created)
the first time. If this information is added during the configuration of
the experiment, it is possible to use the metadata as another part of
the configuration that we merge with `resolve_config.merge_configs()`.

How:

Set datetime when experiment is new and about to register in database.
When experiment is already registered, the datetime will appear in the
configuration fetched from database.