Fix unittests of mongodb and procuder
Why:

The modifications to the experiment database (experiment.yaml)
for the unittests or `ExperimentBuilder` was causing problems in
other submodule unittests. This happened because documents
got added to `experiment.yaml` and changed the indexes of them
in the dictionary definition used in the unittests.

How:

Adapt indexes to the corresponding documents.