Change Trial autogenerated id to a deterministic one
Why: Trial will now require to be unique inside the database. Which
means that the autogenerated ObjectID of MongoDB will not work anymore.
This commit will change how trials' id are calculated by hashing the
parameters representation and the experiment id together as a unique
identifier. This also means that duplicate trials can no longer be
inserted inside the database.