Adapt MongoDB to support options
Why:

`MongoDB` would only pass host, port, username and password to
`MongoClient`. The user could not use any option like `ssl` or
`replicaSet`.

Support for options are mandatory to access secured databases.

How:

Pass the URI rather than the parsed host when specified by the user.

Note:

There is no support for specific options through the interface. Options
need to be specified in the URI for now.

Note:

`pymongo` would overwrite single arguments passed by those in the
URI if they are different. The behavior of `MongoDB` is reversed in this
commit to be coherent with `pymongo`.