Add EphemeralDB
Why:

We need a simple database that does not require any setup for debugging.
This ephemeral db would be built from scratch at every execution of
Oríon, with all the information being lost at the end of the process.

How:

Implement EphemeralDocument, EphemeralCollection and EphemeralDB.

The EphemeralDocument is build using a dictionary, with many methods to
support mongodb-like operations such as selection or query with
operators.

The EphemeralCollection contains a list of EphemeralDocuments and a
dictionary of unique indexes.

Note that the database is not thread-safe. It is not intended to be used
by multiple process anyway. It is created and destroyed within a single
process.