Add --debug option to orion hunt cli
Why:

It is tedious to debug with orion when using a database, especially when
many modifications of invocations keep generating experiment branching events.
With the EphemeralDB, we can isolate invocations and make the debugging
process smoother.

How:

When --debug is specified, the dbtype of the config is overwritten with
'ephemeraldb'