Add base interface for Plotter and Analyser and analysis object
Why: To provide extensibles interfaces for a data analyser and a
plotter, we create base classes DataAnalyser and DataPlotter alongside
their respective factories AnalyserFactory and PlotterFactory. The
actual implementation are wrapped inside the AnalyserWrapper and
PlotterWrapper. Analysis objects serve as connections between an
analyser and a plotter.

Currently, the only analyser and plotter serve as debugger, "analysing"
the best trial and printing it.