Add Concept class for dependency injection classes
Why: Things like `BaseAlgorithm`, `BasePlotter` and the likes all shared
the same code when it came to calling their factory class for
dependency injection and runtime polymorphism. This changes move that
code to another abstract level inside a `Concept` class. This duplicate
code can then be deleted.

How: The `Concept` class deduces some things about the classes
inheriting it: the base class name (`BaseAlgorithm` for example), the
name of the concept (`Algorithm` in that case), and the root module
for the implementations of each subclass.
The `Concept` then creates a factory class from the base class. This
mimics the old behavior of declaring a `ThingFactory` class having
`Factory` as its metaclass (like `OptimisationAlgorithm`).
This change also implies that the factories' name are not available
outside `Concept`. The `Factory` class has been changed as such that
it'll look for the entry points identified by the `Concept` subclass
rather then the `Factory` one.