Add worker-trials argument
Why:
Sometimes it is desirable to limit number of trials to run on a specific node without changing the total number of trials for the whole experiment. For example, one could run only one trial per node on a cluster with max_trials set to some value. This helps to schedule jobs better on a cluster.

How:
--worker-trials affects only a specific worker. The worker exits when either max_trials for the experiment reached or worker_trials.