Adapt core install doc
Why:

Based on issue #132 it looks like the title of the section is
misleading. Also the PyPI link was not added now that we have one.

How:

Change title to `Installation of the Core` and explain how to install the `core`
alone, plus write additional doc on how to install plugins.