Add concurrent unittests for producer
Why:

With new non-blocking design for the generation and registration of new
trials, we need to test that concurrent producers would properly
rollback and update their algorithms when the new trials suggested by
the later are duplicates.

How:

Make sure the algorithms of both producers suggest the same trials, so
that any concurrent generation causes race conditions and force the
second producer to rollback.