Restructure visualization classes
Why: the need to handle different libraries for the plotters led
to some problems along the way. This new structure will now be able
to scale to new libraries easily while being easier to use and
understand. Thee is still some small fixes to be done regarding
the `Analysis` classes but this commit provides a working
visualisation process and tests to go with it.