Keep supporting pymongo<3.7
Why:

Pymongo 3.7 is only a few months old. Bumping the minimal version
requirement to 3.7 to avoid maintaining support for older versions would
be very restrictive. Furthermore, the maintenance cost for supporting
older versions is very low, it is only a try-else cause to use
`count_documents()` is available else `count()`.