Adjust test_stress_unique_folder_creation's max
Why:

The maximum of 1000 trials was busting `max_attempts`. Even though the
limit case of very small learning rates would not be caught by decreased
the number of trials, it is enough to test the creation of unique
folders. The fact that so similar learning rates bust `max_attempts` is
a strong indication that this problem would be reached before the
problem of folder name uniqueness.