Move build_to_config outside of SpaceBuilder
Why: In the spirit of moving parsing and building outside of the
SpaceBuilder, the `build_from` process has been moved to the Orion
specific commandline parser.

As such, the following changes have been made to OrionCmdlineParser:
- The `config` argument now only takes a name of follow the naming
convention of the keys of CmdlineParser.
- Support for `defaultdict` has been added. This means that generic
config file are now supported.
- OrionCmdlineParser will now raise an error of it detects two
arguments with same name inside the config file and the command line.
- A template of the config file is now kept inside `extracted_config`.
- The format function now build the config file to the tmp file given
by the producer as well as adding the `trial` and `experiment` items
to the configuration dictionary.

Fixes:
- A bug where the `~` use inside the value of an argument was
interpreted as a Orion prior has been fixed.
- Config file creation is now working properly.