Relax constraint on script repository
Why:

We decided that Oríon would crash if the script being executed is not
part of a code version control (only support git for now). This was to
ensure reproducibility, but this constraint is troublesome for users
using a different VCS system or none at all. It's not Oríon's role to
enforce a VCS on users. They could at least use Oríon without `git`,
they will only loose the automatic code change detection.

How:

Log a warning that code change cannot be detected instead of raising an error.