Make experiment's user customisable
Why:

The user's name is defined based on the OS's username
(`getpass.getuser()`), but a same user may run the experiment on
different systems using different usernames. There should be a way to
override to default `getpass.getuser()` and pass a username to Oríon to
support such use-cases. This is also true if different users wants to
collaborate and fetch results from a colleague.

How:

Username can be passed in config or as `--user`. The name is then passed
to either `Experiment` or `ExperimentView` in the `ExperimentBuilder`.

Note:

Maybe we should remove the username from the experiment altogether in
the future.