Set default value of pool-size to 1
Why:

Users are often testing sequentially, and even if running in parallel we
now have parallel strategies which makes pool-size > 1 useless unless
the algorithm have strategies different then those implemented in Oríon.
This is such an unlikely scenario that it is better to set the default
value to 1 to avoid confusion.