Set default value of strategy to Max
Why:

Since pool-size is now 1 by default, we want a default strategy that
will handle parallel optimization properly. MaxParallelStrategy is the
most pessimistic one and the one which will mostly favor exploration.
Until we have benchmarkings showing benefits of the different parallel
strategies this ones seems to be the most reasonable one.