Fix multiple bugs inside ASHA
Why: First bug concerns the hash of the points. There was a missing
stringify call to obtain a serie of bytes.
Second bug lies with sorting the current rung inside `get_candidate`.
Since we moved to a `dict` instead of a `set`, we must sort on the
values of the `dict` instead.
Third bug relates to the id of points. Again, since we moved to a
`dict` and are using a hash as an id, we must hash the point to
determine if it is already inside the next rung.
For the sake of consistency, all point are now on unpacked in the
form of `{objective}, {point}`.