Handle bad results in trials
Why:

If the user pass bad results (missing objective, None objective, etc),
Oríon will later crash in mysterious ways for the users. We should
detect this early and raise an insightfull error for the user.

How:

Make the result attribute private and test in results.setter if the
provided results are valid.