Move checks to stages
Why: to be more flexible, the checking process has been separated in
three different stages. The creation of stages was necessary if we
wished to check for default config and environment variables config.

The three stages are the `Presence` stage, which makes sure that at
least one valid configuration exists. The `Creation` stages, which
makes sure that a database can be instantiated. Finally, we have
the `Operations` stage which makes sure all the operations on a
database work.