Use a buffer for experiment.fetch_completed_trials
Why:

If there is an unequal delay between the datetime 'end_time' and the
registration in the database, then _last_fetched in anoter worker might
be more recent then 'end_time', resulting in missed trials in the query.

How:

Buffer trial id and filter based on that. For now we still use
_last_fetched but use `max(fetched_end_times)` instead of `utcnow()` and
also go backward 1 minute before _last_fetched in the query to ensure we
don't miss any trial.