Add PickledDB to support FileSystem backend
Why:

MongoDB does not play nicely with ssh tunnels and we are currently
stucked with a cluster where compute nodes does not have access to
internet, leaving us isolated from MongoDB. Using the file-system is not
optimal but it is at least simpler to set-up for the users.

How:

Simply load a pickled version of EphemeralDB and write the updated
version back to the file. All operations are locked using `filelock` to
provide support for parallel workers.