Skip exp update in ExpView
Why:

The ExperimentView should not update the experiment in the database.
Fields doesn't get updated when loading the View per say, but loading a
view within a different version of Oríon could lead to modification of
the configuration dictionary of the experiment, which would lead to a
modified experiment in the database.

How:

Add `enable_update` argument to `Experiment.configure` so that
`ExperimentView` can call it without updating the DB.

Note:

All this should get refactored once the global and experiment
configuration refactoring is completed.