Bundle update calls in Producer for consistency
Why:

There can be modifications in trials status between the call to
`fetch_completed_trials` and `fetch_noncompleted_trials`, leading
to inconsistent states, where very recently completed trials are neither
in completed trials, nor noncompleted trials. This is problematic for
ASHA, because a promoted trial may be observed by the naive algorithm
while the base trial (non-promoted one) is lost between the 2 calls.
When this happens the naive algo crashes because the rungs are
inconsistent.

How:

Fetch all trials once and then pass the completed ones to
`update_algorithm` and pass the non completed ones to
`update_naive_algorithm`. This way the fetch is a single snapshot and
garantees consistency.