Dont configure experiment views (#223)
Why:

When configuring the experiment, missing local config files can make the
configuarion crash. There is however little value for now to configure
experiment views as we are never using their space or algorithm.
Therefore the simplest solution to this problem is to stop configuring
the experiment views. This may be problematic however for users who may
use the experiment view to fetch and analyse results. For this reason,
we should seek to bring back the full instantiation of experiment views
once the refactoring of orion's global configuarion is done.