Change status --recursive to --collapse
Why:

The name --recursive was confusing for what it was intended to be. The default behavior is to print the EVC tree hierarchically, and --collapse will aggregate results of the EVC in the parent experiment.

Note: (datcorno)

Some fixtures were assigning default values to the x parameter
which lead to conflicts when running `init_only`. Since the `x`
parameter is always inside the parent experiment, the default value
has been removed.

There was a duplicate function that has also been removed.