Move TrialMonitor/TrialPacemaker inside consumer
Why:

We need access to the pacemaker when the trials is completed so that we
can stop it. Otherwise, the thread continues until if makes another
heartbeat before it crashes and quit, which can take lot of time. The
execution of Orion hangs at the end of experiment until the pacemaker
stops.

How:

Move pacemaker inside Consumer so that it can be stopped right at end of
trial execution. Note that there is some delay between the reservation
and the start of pacemaker, which makes it possible to loose the
reservation before execution begins. That is not very likely however
since there is not a large time gap between reservation and execution.