Handle broken refers in experiment configuration
Why:

If the process is stopped between the registration of the experiment and
the update of exp[refers][self._id], then the experiment will be somehow
broken, i.e., many code parts will crash because of it.

How:

Handle refers creation during configuration with default values if
nothing is set so that abortion during registration will in worst case
lead to inconsistant refers that is easier to fix. Previously broken
refers would be {}, which is irrecuperable for child experiments. Now
broken refers would be {parent_id: ok, adapter: ok, root_id: None},
which is simple to infer and fix.