Make Space Sorted instead of Ordered
Why:

The reading of configuration files sometimes return the same data in a
different order (maybe python 3.5 only). Nevertheless, even with a
python API, the user passing a differnt order of hyper-parameter should
not lead to different spaces. Instead of fixing the problem upfront in
the loading of the config, we may fix directly in the space itself and
make it return the dimensions in alphabetical order.