Add expand-versions to status
With the introduction of the `--expand-versions` we now have a new
way of presenting the different experiments to the user. We now
simply display the latest version of the experiment unless one of
the two following conditions is fulfilled:

- Either the user has used the `--expand-versions` argument
- Or, one of the experiment in the version tree has at least one
child named differently.

In both cases all the experiments will be printed.